fix: refactor util functions
