--- src/pages/GiftAid/utils/giftaidDefaults.js
@@ -1,17 +1,0 @@
-/*
-* Common form variables
-*
-*/
-export const hiddenFields = ['field-input--address1', 'field-input--town', 'field-wrapper--country'];
-
-/*
-* REGEX for postcode field
-*
-*/
-export const postCodePattern = '[A-Za-z]{1,2}[0-9Rr][0-9A-Za-z]?( |)[0-9][ABD-HJLNP-UW-Zabd-hjlnp-uw-z]{2}';
-
-/*
-* Just In Time Link Text
-*
-*/
-export const justInTimeLinkText = 'Why do we collect this info?';
